Rosetta: Adjust signing payload (#382)
* Adjust signing payload
- retrieve account public key info in order to know which hashing algorithm to use
- pre-hash signing payload so the signer does not need to know anything about hashing

* Update go mod file

* Function rename

* Rename Rosetta parser to transactor

* Make Flow Go SDK model use explicit

* Make signed and unsigned transactions explicit

* Rename parse field to transact in Rosetta API

* Move key retrieval to invoker

* Moved transaction signing code into transactor
