+++ rosetta/configuration/errors.go
@@ -42,4 +42,5 @@ var (
 	ErrorInvalidReceiver          = meta.ErrorDefinition{Code: 19, Message: "invalid transaction recipient", Retriable: false}
 	ErrorInvalidSignature         = meta.ErrorDefinition{Code: 20, Message: "invalid transaction signature", Retriable: false}
 	ErrorInvalidProposalKey       = meta.ErrorDefinition{Code: 21, Message: "invalid transaction proposal key", Retriable: false}
+	ErrorInvalidAuthorizerKey     = meta.ErrorDefinition{Code: 22, Message: "invalid transaction authorizer key", Retriable: false}
 )
