+++ rosetta/transactor/transactor.go
@@ -1,0 +1,487 @@
+// Copyright 2021 Optakt Labs OÜ
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not
+// use this file except in compliance with the License. You may obtain a copy of
+// the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations under
+// the License.
+
+package transactor
+
+import (
+	"bytes"
+	"encoding/hex"
+	"fmt"
+	"sort"
+	"strconv"
+
+	"github.com/onflow/cadence"
+	"github.com/onflow/cadence/encoding/json"
+	"github.com/onflow/flow-go-sdk/crypto"
+
+	sdk "github.com/onflow/flow-go-sdk"
+	"github.com/onflow/flow-go/model/flow"
+
+	"github.com/optakt/flow-dps/models/dps"
+	"github.com/optakt/flow-dps/rosetta/failure"
+	"github.com/optakt/flow-dps/rosetta/identifier"
+	"github.com/optakt/flow-dps/rosetta/object"
+)
+
+const (
+	// Transactions should have exactly one authorizer.
+	authorizersRequired = 1
+
+	// Transaction script should have exactly two arguments.
+	argsRequired = 2
+)
+
+// Transactor has the capabilities to determine transaction intent from an array
+// of Rosetta operations, create a Flow transaction from a transaction intent
+// and transate a Flow transaction back to an array of Rosetta operations.
+type Transactor struct {
+	validate Validator
+	generate Generator
+	invoke   Invoker
+}
+
+// NewTransactor creates a new transaction Transactor to handle constructing
+// and parsing transactions.
+func New(validate Validator, generate Generator, invoke Invoker) *Transactor {
+
+	p := Transactor{
+		validate: validate,
+		generate: generate,
+		invoke:   invoke,
+	}
+
+	return &p
+}
+
+// DeriveIntent derives a transaction Intent from two operations given as input.
+// Specified operations should be symmetrical, a deposit and a withdrawal from two
+// different accounts. At the moment, the only fields taken into account are the
+// account IDs, amounts and type of operation.
+func (t *Transactor) DeriveIntent(operations []object.Operation) (*Intent, error) {
+
+	// Verify that we have exactly two operations.
+	if len(operations) != 2 {
+		return nil, failure.InvalidOperations{
+			Description: failure.NewDescription("invalid number of operations"),
+			Count:       len(operations),
+		}
+	}
+
+	amounts := make(map[int]int64)
+
+	// Parse amounts.
+	for i, op := range operations {
+		amount, err := strconv.ParseInt(op.Amount.Value, 10, 64)
+		if err != nil {
+			return nil, failure.InvalidIntent{
+				Description: failure.NewDescription("could not parse amount",
+					failure.WithString("amount", op.Amount.Value),
+					failure.WithErr(err),
+				),
+			}
+		}
+
+		amounts[i] = amount
+	}
+
+	// Verify that the amounts match.
+	if amounts[0] != -amounts[1] {
+		return nil, failure.InvalidIntent{
+			Description: failure.NewDescription("transfer amounts do not match",
+				failure.WithString("first_amount", operations[0].Amount.Value),
+				failure.WithString("second_amount", operations[1].Amount.Value),
+			),
+		}
+	}
+
+	// Sort the operations so that the send operation (negative amount) comes first.
+	sort.Slice(operations, func(i, j int) bool {
+		return amounts[i] < amounts[j]
+	})
+
+	send := operations[0]
+	receive := operations[1]
+
+	// Validate the currencies specified for deposit and withdrawal.
+	sendSymbol, _, err := t.validate.Currency(send.Amount.Currency)
+	if err != nil {
+		return nil, fmt.Errorf("invalid sender currency: %w", err)
+	}
+	receiveSymbol, _, err := t.validate.Currency(receive.Amount.Currency)
+	if err != nil {
+		return nil, fmt.Errorf("invalid receiver currency: %w", err)
+	}
+
+	// Make sure that both the send and receive operations are for FLOW tokens.
+	if sendSymbol != dps.FlowSymbol || receiveSymbol != dps.FlowSymbol {
+
+		return nil, failure.InvalidIntent{
+			Description: failure.NewDescription("invalid currencies found",
+				failure.WithString("sender", send.AccountID.Address),
+				failure.WithString("receiver", receive.AccountID.Address),
+				failure.WithString("withdrawal_currency", send.Amount.Currency.Symbol),
+				failure.WithString("deposit_currency", receive.Amount.Currency.Symbol)),
+		}
+	}
+
+	// Validate the sender and the receiver account IDs.
+	_, err = t.validate.Account(send.AccountID)
+	if err != nil {
+		return nil, fmt.Errorf("invalid sender account: %w", err)
+	}
+	_, err = t.validate.Account(receive.AccountID)
+	if err != nil {
+		return nil, fmt.Errorf("invalid receiver account: %w", err)
+	}
+
+	// Validate that the specified operations are transfers.
+	if send.Type != dps.OperationTransfer || receive.Type != dps.OperationTransfer {
+		return nil, failure.InvalidIntent{
+			Description: failure.NewDescription("only transfer operations are supported",
+				failure.WithString("withdrawal_type", send.Type),
+				failure.WithString("deposit_type", receive.Type),
+			),
+		}
+	}
+
+	amount := amounts[0]
+	if amount < 0 {
+		amount = -amount
+	}
+
+	intent := Intent{
+		From:     flow.HexToAddress(send.AccountID.Address),
+		To:       flow.HexToAddress(receive.AccountID.Address),
+		Amount:   cadence.UFix64(amount),
+		Payer:    flow.HexToAddress(send.AccountID.Address),
+		Proposer: flow.HexToAddress(send.AccountID.Address),
+		GasLimit: flow.DefaultMaxTransactionGasLimit,
+	}
+
+	return &intent, nil
+}
+
+// CompileTransaction creates a complete Flow transaction from the given intent and metadata.
+func (t *Transactor) CompileTransaction(refBlockID identifier.Block, intent *Intent, sequence uint64) (*sdk.Transaction, error) {
+
+	// Generate script for the token transfer.
+	script, err := t.generate.TransferTokens(dps.FlowSymbol)
+	if err != nil {
+		return nil, fmt.Errorf("could not generate transfer script: %w", err)
+	}
+
+	// TODO: Allow arbitrary proposal key index
+	// => https://github.com/optakt/flow-dps/issues/369
+
+	// Create the transaction.
+	tx := sdk.NewTransaction().
+		SetScript(script).
+		SetReferenceBlockID(sdk.HexToID(refBlockID.Hash)).
+		SetPayer(sdk.Address(intent.Payer)).
+		SetProposalKey(sdk.Address(intent.Proposer), 0, sequence).
+		AddAuthorizer(sdk.Address(intent.From)).
+		SetGasLimit(intent.GasLimit)
+
+	receiver := cadence.NewAddress(flow.BytesToAddress(intent.To.Bytes()))
+
+	// Add the script arguments - the amount and the receiver.
+	_ = tx.AddArgument(intent.Amount)
+	_ = tx.AddArgument(receiver)
+
+	return tx, nil
+}
+
+func (t *Transactor) HashPayload(rosBlockID identifier.Block, unsignedTx *sdk.Transaction, signer identifier.Account) ([]byte, error) {
+
+	// Validate block.
+	height, _, err := t.validate.Block(rosBlockID)
+	if err != nil {
+		return nil, fmt.Errorf("could not validate block: %w", err)
+	}
+
+	// Validate address.
+	address, err := t.validate.Account(signer)
+	if err != nil {
+		return nil, fmt.Errorf("could not validate account: %w", err)
+	}
+
+	key, err := t.invoke.Key(height, address, 0)
+	if err != nil {
+		return nil, fmt.Errorf("could not retrieve key: %w", err)
+	}
+
+	message := unsignedTx.EnvelopeMessage()
+	message = append(flow.TransactionDomainTag[:], message...)
+
+	hasher, err := crypto.NewHasher(key.HashAlgo)
+	if err != nil {
+		return nil, fmt.Errorf("could not retrieve hasher: %w", err)
+	}
+
+	hash := hasher.ComputeHash(message)
+
+	return hash, nil
+}
+
+// ParseTransactions processes the flow transaction, validates its correctness and translates it
+// to a list of operations and a list of signers.
+func (t *Transactor) ParseTransaction(tx *sdk.Transaction) ([]object.Operation, []identifier.Account, error) {
+
+	// Validate the transaction actors. We expect a single authorizer - the sender account.
+	// For now, the sender must also be the proposer and the payer for the transaction.
+
+	if len(tx.Authorizers) != authorizersRequired {
+		return nil, nil, failure.InvalidAuthorizers{
+			Have:        uint(len(tx.Authorizers)),
+			Want:        authorizersRequired,
+			Description: failure.NewDescription("invalid number of authorizers"),
+		}
+	}
+
+	authorizer := tx.Authorizers[0]
+	sender := identifier.Account{
+		Address: authorizer.String(),
+	}
+
+	// Validate the sender address.
+	_, err := t.validate.Account(sender)
+	if err != nil {
+		return nil, nil, fmt.Errorf("invalid sender account: %w", err)
+	}
+
+	// Verify that the sender is the payer and the proposer.
+	if tx.Payer != authorizer {
+		return nil, nil, failure.InvalidPayer{
+			Have:        convertAddress(tx.Payer),
+			Want:        convertAddress(authorizer),
+			Description: failure.NewDescription("invalid transaction payer"),
+		}
+	}
+	if tx.ProposalKey.Address != authorizer {
+		return nil, nil, failure.InvalidProposer{
+			Have:        convertAddress(tx.ProposalKey.Address),
+			Want:        convertAddress(authorizer),
+			Description: failure.NewDescription("invalid transaction proposer"),
+		}
+	}
+
+	// Verify the transaction script is the token transfer script.
+	script, err := t.generate.TransferTokens(dps.FlowSymbol)
+	if err != nil {
+		return nil, nil, fmt.Errorf("could not generate transfer script: %w", err)
+	}
+	if !bytes.Equal(script, tx.Script) {
+		return nil, nil, failure.InvalidScript{
+			Script:      string(tx.Script),
+			Description: failure.NewDescription("transaction text is not valid token transfer script"),
+		}
+	}
+
+	// Verify that the transaction script has the correct number of arguments.
+	args := tx.Arguments
+	if len(args) != argsRequired {
+		return nil, nil, failure.InvalidArguments{
+			Have:        uint(len(args)),
+			Want:        argsRequired,
+			Description: failure.NewDescription("invalid number of arguments"),
+		}
+	}
+
+	// Parse and validate the amount argument.
+	val, err := json.Decode(args[0])
+	if err != nil {
+		return nil, nil, failure.InvalidAmount{
+			Amount: string(args[0]),
+			Description: failure.NewDescription("could not parse transaction amount",
+				failure.WithErr(err)),
+		}
+	}
+	amountArg, ok := val.ToGoValue().(uint64)
+	if !ok {
+		return nil, nil, failure.InvalidAmount{
+			Amount:      string(args[0]),
+			Description: failure.NewDescription("invalid amount"),
+		}
+	}
+	amount := strconv.FormatUint(amountArg, 10)
+
+	// Parse and validate receiver script argument.
+	val, err = json.Decode(args[1])
+	if err != nil {
+		return nil, nil, failure.InvalidReceiver{
+			Receiver: string(args[1]),
+			Description: failure.NewDescription("could not parse transaction receiver address",
+				failure.WithErr(err)),
+		}
+	}
+	addr := flow.HexToAddress(val.String())
+	receiver := identifier.Account{
+		Address: addr.String(),
+	}
+	_, err = t.validate.Account(receiver)
+	if err != nil {
+		return nil, nil, fmt.Errorf("invalid receiver account: %w", err)
+	}
+
+	// Validate the reference block identifier.
+	rosBlockID := identifier.Block{
+		Hash: tx.ReferenceBlockID.String(),
+	}
+	_, _, err = t.validate.Block(rosBlockID)
+	if err != nil {
+		return nil, nil, fmt.Errorf("invalid reference block: %w", err)
+	}
+
+	// Create the send operation.
+	sendOp := object.Operation{
+		ID: identifier.Operation{
+			Index: 0,
+		},
+		AccountID: sender,
+		Type:      dps.OperationTransfer,
+		Amount: object.Amount{
+			Value: "-" + amount,
+			Currency: identifier.Currency{
+				Symbol:   dps.FlowSymbol,
+				Decimals: dps.FlowDecimals,
+			},
+		},
+	}
+
+	// Create the receive operation.
+	receiveOp := object.Operation{
+		ID: identifier.Operation{
+			Index: 1,
+		},
+		AccountID: receiver,
+		Type:      dps.OperationTransfer,
+		Amount: object.Amount{
+			Value: amount,
+			Currency: identifier.Currency{
+				Symbol:   dps.FlowSymbol,
+				Decimals: dps.FlowDecimals,
+			},
+		},
+	}
+
+	// Create the operations list.
+	ops := []object.Operation{
+		sendOp,
+		receiveOp,
+	}
+
+	// Since we only support sender as the payer/proposer, we never expect any payload signatures.
+	if len(tx.PayloadSignatures) > 0 {
+		return nil, nil, failure.InvalidSignature{
+			Description: failure.NewDescription("unexpected payload signature found",
+				failure.WithInt("signatures", len(tx.PayloadSignatures))),
+		}
+	}
+
+	// We may be parsing an unsigned transaction - if that's the case, we're done.
+	if len(tx.EnvelopeSignatures) == 0 {
+		return ops, nil, nil
+	}
+
+	// We don't support multiple signatures.
+	if len(tx.EnvelopeSignatures) > 1 {
+		return nil, nil, failure.InvalidSignature{
+			Description: failure.NewDescription("unexpected envelope signatures found",
+				failure.WithInt("signatures", len(tx.EnvelopeSignatures))),
+		}
+	}
+
+	// Validate that it is the sender who signed the transaction.
+	signer := tx.EnvelopeSignatures[0].Address
+	if signer != authorizer {
+		return nil, nil, failure.InvalidSignature{
+			Description: failure.NewDescription("invalid signer account",
+				failure.WithString("have_signer", signer.String()),
+				failure.WithString("want_signer", authorizer.String()),
+				failure.WithString("signature", hex.EncodeToString(tx.EnvelopeSignatures[0].Signature))),
+		}
+	}
+
+	// TODO: Validate the actual signature.
+	// => https://github.com/optakt/flow-dps/issues/376
+
+	// Create the signers list.
+	signers := []identifier.Account{
+		sender,
+	}
+
+	return ops, signers, nil
+}
+
+// AttachSignature adds the given signature to the transaction.
+func (t *Transactor) AttachSignature(tx *sdk.Transaction, signature object.Signature) (*sdk.Transaction, error) {
+
+	// Validate the transaction actors. We expect a single authorizer - the sender account.
+	if len(tx.Authorizers) != authorizersRequired {
+		return nil, failure.InvalidAuthorizers{
+			Have:        uint(len(tx.Authorizers)),
+			Want:        authorizersRequired,
+			Description: failure.NewDescription("invalid number of authorizers"),
+		}
+	}
+
+	sender := tx.Authorizers[0]
+
+	// Verify that the sender is the payer, since it is the payer that needs to sign the envelope.
+	if tx.Payer != sender {
+		return nil, failure.InvalidPayer{
+			Have:        convertAddress(tx.Payer),
+			Want:        convertAddress(sender),
+			Description: failure.NewDescription("invalid transaction payer"),
+		}
+	}
+
+	// Verify that we do not already have signatures.
+	if len(tx.EnvelopeSignatures) > 0 {
+		return nil, failure.InvalidSignature{
+			Description: failure.NewDescription("unexpected envelope signatures found",
+				failure.WithInt("signatures", len(tx.EnvelopeSignatures))),
+		}
+	}
+
+	// Verify that the signature belongs to the sender.
+	signer := sdk.HexToAddress(signature.SigningPayload.AccountID.Address)
+	if signer != sender {
+		return nil, failure.InvalidSignature{
+			Description: failure.NewDescription("invalid signer account",
+				failure.WithString("have_signer", signer.String()),
+				failure.WithString("want_signer", sender.String())),
+		}
+	}
+
+	bytes, err := hex.DecodeString(signature.HexBytes)
+	if err != nil {
+		return nil, failure.InvalidSignature{
+			Description: failure.NewDescription("invalid signature payload",
+				failure.WithErr(err)),
+		}
+	}
+
+	// Copy the transaction and add the signature.
+	signedTx, err := sdk.DecodeTransaction(tx.Encode())
+	if err != nil {
+		return nil, fmt.Errorf("could not copy transaction: %w", err)
+	}
+
+	// TODO: allow arbitrary key index
+	// => https://github.com/optakt/flow-dps/issues/369
+	signedTx.AddEnvelopeSignature(signer, 0, bytes)
+
+	return signedTx, nil
+}
