+++ rosetta/transactor/validator.go
@@ -12,7 +12,7 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-package transactions
+package transactor
 
 import (
 	"github.com/onflow/flow-go/model/flow"
